backend/requirements.txt
flask
python-dotenv
flask-restful
flask_sqlalchemy
flask-jwt-extended

